+++ ChangeLog.md
@@ -3,6 +3,7 @@
 
 ## Unreleased
 
+* NEW Add option SELLYOURSAAS_FORCE_NO_SELECTION_IF_SEVERAL
 * NEW More protection on public registration
 * NEW Can show the label on server on registration page
 * FIX Add alter table for column 'label' on llx_sellyoursaas_deploymentserver